+++ src/main/java/org/tugraz/sysds/runtime/instructions/cp/UnaryScalarCPInstruction.java
@@ -65,7 +65,7 @@ public class UnaryScalarCPInstruction extends UnaryMatrixCPInstruction {
 		}
 		else {
 			UnaryOperator dop = (UnaryOperator) _optr;
-			if ( so instanceof IntObject && output.getValueType() == ValueType.INT )
+			if ( so instanceof IntObject && output.getValueType() == ValueType.INT64 )
 				sores = new IntObject((long)dop.fn.execute(so.getLongValue()));
 			else if( so instanceof BooleanObject && output.getValueType() == ValueType.BOOLEAN )
 				sores = new BooleanObject(dop.fn.execute(so.getBooleanValue()));
